Eliminating logic to cope with 'ifconfig' particular approach.
By eliminating this logic we are enabling the user to configure overlapping subnets within the same interface. This is a requirement for link-local scenarios in which more than one link-local subnet (fe80/10) may be configured within the same interface. In a subsequent commit the 'ifconfig' utility will be eliminated from the sonic image.

Please refer to this link for details about the testing being done: https://iwww.corp.linkedin.com/wiki/cf/display/IEO/Link-Local+forwarding

RB=1282553
G=lnos-reviewers
R=ntrianta,pmao,rmolina,sfardeen,zxu
A=