Fix CDP/UDLD etc Cisco packet being shown as rx discards
On L3 links connected to Cisco NX devices where CDP/UDLD protocol was enabled,
we would see rx discards in the "show interface counters" output.

The reason being that the CDP/UDLD packets were not treated specifically on HW,
thus it was considered as unknown multicast packets.  In case, the ingress port
is a L3 port, it has a reserved vlan on it, and it has unknown mcast flood enabled.
Since the vlan only has one port in it, when it flood it could not find outgoing port,
and thus cause rx discard on BRCM platform.

The fix is to use UDLD trap (the only one in SAI) and sai implementation was
to use the Cisco multicast DMAC (01-00-0c-cc-cc-cc) to trap the packets.

Priority 1 is used since we don't handle UDLD for now,  so it is not competing
with priority 4 traffic.

SAI PR was sent to add a more generic trap rather than UDLD.

    RB=1313473
    G=lnos-reviewers
    R=pchaudha,pmao,rmolina,zxu
    A=