ssh is requested to be prioritied in case DDOS or other network storm.
It is placed below bgp/lldp etc protocols, so it won't be competing with
them in case of scp/sftp. But it is higher than other groups.

Rate limit is also set to protect the case when you scp/sftp traffic.
We should get ~4.5MB/s in case of scp/sftp with MTU = 1500.

    RB=1320954
    G=lnos-reviewers
    R=pchaudha,pmao,rmolina,zxu
    A=