Support VLAN auto-state off feature.
In today's code, if no vlan port is up in a vlan, the vlan
interface would go down. In Linkedin deployment, we wanted to
keep vlan interface always up so we can access to the device
through vlan interface even if all ports in the vlan are down,
e,g, servers are not populated in a POD.

The changes added a dummy interface to bridge, and keep the dummy
interface always up so the vlan interface would be up too.
Also added a logic to handle the dummy interface netlink messages
differently.

Signed-off-by: Zhenggen Xu <zxu@linkedin.com>

    RB=1522986
    G=lnos-reviewers
    R=pchaudha,pmao,rmolina,zxu
    A=